+++ ewah/dag_factories/dag_factory_idempotent.py
@@ -4,7 +4,7 @@ from airflow.operators.bash import BashOperator
 
 from ewah.constants import EWAHConstants as EC
 from ewah.uploaders.snowflake import SnowflakeOperator
-from ewah.ewah_utils.airflow_utils import PGO, etl_schema_tasks, datetime_utcnow_with_tz
+from ewah.utils.airflow_utils import PGO, etl_schema_tasks, datetime_utcnow_with_tz
 from ewah.hooks.base import EWAHBaseHook as BaseHook
 from ewah.operators.base import EWAHBaseOperator
