+++ ewah/hooks/google_ads.py
@@ -38,7 +38,7 @@ class EWAHGoogleAdsHook(EWAHBaseHook):
     def get_connection_form_widgets() -> dict:
         """Returns connection widgets to add to connection form"""
         # from flask_appbuilder.fieldwidgets import BS3TextFieldWidget
-        from ewah.ewah_utils.widgets import EWAHTextAreaWidget
+        from ewah.utils.widgets import EWAHTextAreaWidget
         from wtforms import PasswordField
 
         return {
