+++ ewah/operators/fx.py
@@ -3,7 +3,7 @@ from yahoofinancials import YahooFinancials
 
 from ewah.constants import EWAHConstants as EC
 from ewah.operators.base import EWAHBaseOperator
-from ewah.ewah_utils.airflow_utils import datetime_utcnow_with_tz
+from ewah.utils.airflow_utils import datetime_utcnow_with_tz
 
 
 class EWAHFXOperator(EWAHBaseOperator):
