+++ ewah/operators/shopify.py
@@ -1,5 +1,5 @@
 from ewah.operators.base import EWAHBaseOperator
-from ewah.ewah_utils.python_utils import is_iterable_not_string
+from ewah.utils.python_utils import is_iterable_not_string
 from ewah.constants import EWAHConstants as EC
 
 from ewah.hooks.base import EWAHBaseHook as BaseHook
