+++ setup.cfg
@@ -3,4 +3,4 @@ description-file = README.md
 
 [options.entry_points]
 apache_airflow_provider=
-    provider_info=ewah.ewah_utils.airflow_provider_info:get_provider_info
+    provider_info=ewah.utils.airflow_provider_info:get_provider_info
